requirements.txt
python-dotenv>=1.0.1
pydantic>=2.6.0
tenacity>=8.2.3
streamlit>=1.32.0
openai>=1.0.0
